requirements.txt
slack_bolt
slack_sdk
openai
python-dotenv
